Lambda + batch mode
